Tattoos
First draft. Added a TattooClass and vars of tattoo class types to Creature class.

Different tattoo types get own classes to set flags and functions.  Rather than put massive nested switch case structures in apperance, used function that has creature target as param to get description of tattoos.

It would be difficult to set a single description for a tattoo as the players body can transform and often times transformations occur in many places so passing the creature as a param allows the tattoos description to alter based on the current body types.

Using Kinky Inks Inc docs as means for player to apply tattoos to themselves.

No removal as of yet